LPS-80681 Rename error module to exception-mapper-impl
