--- modules/apps/apio-architect/apio-architect-error/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Apio Architect Error
-Bundle-SymbolicName: com.liferay.apio.architect.error
-Bundle-Version: 1.0.0
