+++ modules/apps/apio-architect/apio-architect-exception-mapper-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Apio Architect Exception Mapper Impl
+Bundle-SymbolicName: com.liferay.apio.architect.exception.mapper.impl
+Bundle-Version: 1.0.0
